+++ src/packer/engine/pack/__init__.py
@@ -1,1 +1,2 @@
+from packer.engine.pack import arch as _arch  # noqa: F401  (self-registration)
 from packer.engine.pack import tokenizer as _tokenizer  # noqa: F401  (self-registration)